Rename TaskThrottler to Throttler
